+++ spec/ultra_settings/configuration_view_spec.rb
@@ -3,14 +3,9 @@
 require_relative "../spec_helper"
 
 describe UltraSettings::ConfigurationView do
-  it "renders the confguration as an HTML table" do
+  it "renders the configuration as HTML" do
     html = UltraSettings::ConfigurationView.new(TestConfiguration.instance).render
-    expect(html.strip).to match(/\A<table class="ultra-settings-table">.*<\/table>\z/m)
-  end
-
-  it "can set the table class" do
-    html = UltraSettings::ConfigurationView.new(TestConfiguration.instance).render(table_class: "table table-striped")
-    expect(html.strip).to match(/\A<table class="table table-striped">.*<\/table>\z/m)
+    expect(html.strip).to match(/\A.*ultra-settings-fields.*\z/m)
   end
 
   it "escapes HTML in the configuration values", env: {TEST_STRING: "<script>"} do