+++ spec/ultra_settings_spec.rb
@@ -50,17 +50,22 @@ describe UltraSettings do
 
     it "returns nil if not set" do
       UltraSettings.runtime_settings_url = nil
-      expect(UltraSettings.runtime_settings_url("foo")).to be_nil
+      expect(UltraSettings.runtime_settings_url("foo", "string")).to be_nil
     end
 
     it "return nil if set to empty" do
       UltraSettings.runtime_settings_url = ""
-      expect(UltraSettings.runtime_settings_url("foo")).to be_nil
+      expect(UltraSettings.runtime_settings_url("foo", "string")).to be_nil
     end
 
     it "returns the url with the ${name} placeholder replaced with the name argument" do
       UltraSettings.runtime_settings_url = "http://example.com/settings?filter=${name}"
-      expect(UltraSettings.runtime_settings_url("foo bar")).to eq("http://example.com/settings?filter=foo+bar")
+      expect(UltraSettings.runtime_settings_url("foo bar", "string")).to eq("http://example.com/settings?filter=foo+bar")
+    end
+
+    it "returns the url with the ${type} placeholder replaced with the type argument" do
+      UltraSettings.runtime_settings_url = "http://example.com/settings?filter=${type}"
+      expect(UltraSettings.runtime_settings_url("foo bar", "string")).to eq("http://example.com/settings?filter=string")
     end
   end
 end